apps/api/requirements/requirements.txt
accelerate==1.8.1
einops==0.8.1
ftfy==6.3.1
gguf==0.17.1
hf-transfer==0.1.9
huggingface-hub==0.35.3
imageio==2.37.0
loguru==0.7.3
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
peft==0.17.0
pillow==11.2.1
protobuf==6.31.1
pydash==8.0.5
pyyaml==6.0.2
jsonschema>=4.21
requests==2.32.4
safetensors==0.5.3
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.57.1
urllib3==2.5.0
google-cloud-storage
azure-storage-blob
boto3
tenacity
tiktoken
blobfile
sentencepiece
imageio-ffmpeg
timm
scipy
scikit-image
matplotlib
onnxruntime
librosa
pyloudnorm
ffmpeg-python
pip==22.3.1
setuptools==75.8.2
gdown
psutil
gunicorn==23.0.0
ray[default,dashboard]==2.40.0
aiohttp
aiohttp-cors
grpcio
honcho==2.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
python-multipart==0.0.20
av==16.0.1
dotenv
ujson
remap
boltons